GN:restructure service 4, ready to implement db
