+++ magic8ball/requirements.txt
@@ -1,4 +1,5 @@
 flask
+flask-sqlalchemy
 flask_testing
 pytest
 pytest-cov
