+++ magic8ball/service_1/templates/base.html
@@ -1,0 +1,7 @@
+<center>
+    <h1>Welcome!</h1>
+<h2>The MAGIC 8 BALL sees you!</h2>
+<h4>Ask your question and spam refresh!!!</h4>
+</center>
+{% block body %}
+{% endblock body %}
