+++ magic8ball/service_1/templates/home.html
@@ -1,0 +1,17 @@
+{% extends 'base.html' %}
+
+{% block body %}
+<div class='navbar'>
+    How did we get here? service_2 sent {{ a }} and service_3 sent {{ b }}. 
+    Service_4 multiplied them and got {{ x.m }}. See the prophecy related below!
+</div>
+<div class='prophecy'>
+    <center>
+        <h4>
+            The MAGIC 8 BALL says: <br>
+            {{ x.prophecy }}
+        </h4>
+    </center>
+</div>
+
+{% endblock body %}
