+++ magic8ball/service_4/application/__init__.py
@@ -1,0 +1,13 @@
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+import os
+
+app = Flask(__name__)
+
+app.config['SQLALCHEMY_DATABASE_URI'] = str(os.getenv('DATABASE_URI'))  #change this to mysqlserver
+app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+app.config['SECRET_KEY'] = str(os.urandom(16))
+
+db = SQLAlchemy(app)
+
+from application import routes
