+++ magic8ball/service_4/application/models.py
@@ -1,0 +1,5 @@
+from application import db
+
+class Nexus(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    omen = db.Column(db.String(200), nullable=False)
